refactor: 역할에 맞게 서비스 통합
