+++ src/test/java/roomescape/service/ThemeServiceTest.java
@@ -9,6 +9,9 @@ import roomescape.domain.dto.ThemeRequest;
 import roomescape.domain.dto.ThemeResponse;
 import roomescape.exception.DeleteNotAllowException;
 
+import java.time.LocalDate;
+import java.util.List;
+
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -68,5 +71,17 @@ class ThemeServiceTest {
         assertThatThrownBy(() -> service.delete(1L)).isInstanceOf(DeleteNotAllowException.class);
         assertThat(getThemeSize()).isEqualTo(initialSize);
 
+    }
+
+    @DisplayName("기간이 주어지면 가장 많이 예약한 테마 목록 순으로 조회 결과가 반환된다.")
+    @Test
+    void givenStartDateEndDateCount_when_getPopularThemeListAndGetFirst_then_getMostReservedTheme() {
+        //given
+        LocalDate startDate = LocalDate.parse("2024-04-30");
+        LocalDate endDate = LocalDate.parse("2024-05-02");
+        Long count = 10L;
+        //when
+        final List<ThemeResponse> popularThemeList = service.getPopularThemeList(startDate, endDate, count);
+        assertThat(popularThemeList.get(0).id()).isEqualTo(2L);
     }
 }
